Extract shared DNS name-poisoning mixin
Move the selective DNS poisoning behaviour out of the DHCPv6 DNS takeover
module into a reusable DNS::NamePoisoner mixin (TARGET_DOMAIN, TARGET_HOSTS,
SPOOF_IP6, RELAY_CNAME plus the on_dispatch_request poison/forward logic and
the IPv6 validation helper), so the upcoming Router Advertisement takeover
module can share the exact same coercion behaviour instead of duplicating it.

The DHCPv6 module now just includes the mixin and keeps its DHCPv6-specific
options and server wiring. No behaviour change; its spec is unchanged and green.
